Removed old template bot code
