+++ Finished/PictureBot/Dialogs/OrderDialog.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace PictureBot.Dialogs
+{
+    public enum SizeOptions
+    {
+        EightByTen,
+        FiveBySeven,
+        Wallet
+    };
+
+    public enum FinishOptions
+    {
+        Glossy,
+        Matte
+    };
+
+
+    [Serializable]
+    public class OrderDialog
+    {
+    }
+}
