+++ Finished/PictureBot/Dialogs/SearchDialog.cs
