+++ Finished/PictureBot/Models/Image.cs
