+++ Finished/PictureBot/Models/ImageMapper.cs
