+++ Finished/PictureBot/Models/SearchHit.cs
