+++ Finished/PictureBot/Web.config
@@ -9,6 +9,10 @@
     <add key="BotId" value="YourBotId" />
     <add key="MicrosoftAppId" value="" />
     <add key="MicrosoftAppPassword" value="" />
+    <!-- Azure Search Settings -->
+    <add key="SearchDialogsServiceName" value="" />
+    <add key="SearchDialogsServiceKey" value="" />
+    <add key="SearchDialogsIndexName" value="images" />
   </appSettings>
   <!--
     For a description of web.config changes see http://go.microsoft.com/fwlink/?LinkId=235367.
