--- LUIS/CognitiveServicesTutorialLuisModel.json
@@ -1,279 +1,0 @@
-{
-  "luis_schema_version": "2.0.0",
-  "versionId": "0.1",
-  "name": "CognitiveServicesTutorialLuisModel",
-  "desc": "This LUIS model is used with the Cognitive Services tutorial located at https://github.com/noodlefrenzy/CognitiveServicesTutorial.  ",
-  "culture": "en-us",
-  "intents": [
-    {
-      "name": "Greeting"
-    },
-    {
-      "name": "None"
-    },
-    {
-      "name": "SearchPics"
-    },
-    {
-      "name": "SharePic"
-    }
-  ],
-  "entities": [
-    {
-      "name": "facet"
-    }
-  ],
-  "composites": [],
-  "closedLists": [
-    {
-      "name": "emotion",
-      "subLists": [
-        {
-          "canonicalForm": "sadness",
-          "list": [
-            "sad",
-            "unhappy",
-            "miserable"
-          ]
-        },
-        {
-          "canonicalForm": "anger",
-          "list": [
-            "angry",
-            "mad",
-            "enraged"
-          ]
-        },
-        {
-          "canonicalForm": "contempt",
-          "list": [
-            "contemptuous"
-          ]
-        },
-        {
-          "canonicalForm": "disgust",
-          "list": [
-            "disgusted"
-          ]
-        },
-        {
-          "canonicalForm": "fear",
-          "list": [
-            "afraid",
-            "scared",
-            "fearful"
-          ]
-        },
-        {
-          "canonicalForm": "happiness",
-          "list": [
-            "happy",
-            "joyful",
-            "ecstatic"
-          ]
-        },
-        {
-          "canonicalForm": "neutral",
-          "list": [
-            "bored"
-          ]
-        },
-        {
-          "canonicalForm": "surprise",
-          "list": [
-            "surprised",
-            "shocked"
-          ]
-        }
-      ]
-    },
-    {
-      "name": "gender",
-      "subLists": [
-        {
-          "canonicalForm": "male",
-          "list": [
-            "men",
-            "man",
-            "boy",
-            "boys",
-            "guy",
-            "guys"
-          ]
-        },
-        {
-          "canonicalForm": "female",
-          "list": [
-            "women",
-            "woman",
-            "girls",
-            "girl",
-            "ladies",
-            "lady"
-          ]
-        }
-      ]
-    }
-  ],
-  "bing_entities": [
-    "age"
-  ],
-  "actions": [],
-  "model_features": [],
-  "regex_features": [],
-  "utterances": [
-    {
-      "text": "do you like peanut butter and jelly?",
-      "intent": "None",
-      "entities": []
-    },
-    {
-      "text": "reset",
-      "intent": "None",
-      "entities": []
-    },
-    {
-      "text": "echo: i love you",
-      "intent": "None",
-      "entities": []
-    },
-    {
-      "text": "tweet",
-      "intent": "SharePic",
-      "entities": []
-    },
-    {
-      "text": "share",
-      "intent": "SharePic",
-      "entities": []
-    },
-    {
-      "text": "share that one",
-      "intent": "SharePic",
-      "entities": []
-    },
-    {
-      "text": "tweet that picture",
-      "intent": "SharePic",
-      "entities": []
-    },
-    {
-      "text": "share this pic",
-      "intent": "SharePic",
-      "entities": []
-    },
-    {
-      "text": "tweet this picture",
-      "intent": "SharePic",
-      "entities": []
-    },
-    {
-      "text": "find pictures of women outdoors",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "show me happy boy pics",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "i want to see pics of sad women",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "show me pictures of boys looking sad",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "search for pictures of women indoors",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "i want to find outdoor photos",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "show me food pics",
-      "intent": "SearchPics",
-      "entities": [
-        {
-          "entity": "facet",
-          "startPos": 8,
-          "endPos": 11
-        }
-      ]
-    },
-    {
-      "text": "find pictures of food",
-      "intent": "SearchPics",
-      "entities": [
-        {
-          "entity": "facet",
-          "startPos": 17,
-          "endPos": 20
-        }
-      ]
-    },
-    {
-      "text": "please give me pics of 20 year old women",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "find photos of 6-month-old boys",
-      "intent": "SearchPics",
-      "entities": []
-    },
-    {
-      "text": "find pictures of children",
-      "intent": "SearchPics",
-      "entities": [
-        {
-          "entity": "facet",
-          "startPos": 17,
-          "endPos": 24
-        }
-      ]
-    },
-    {
-      "text": "whazzup",
-      "intent": "Greeting",
-      "entities": []
-    },
-    {
-      "text": "yo",
-      "intent": "Greeting",
-      "entities": []
-    },
-    {
-      "text": "hey bot",
-      "intent": "Greeting",
-      "entities": []
-    },
-    {
-      "text": "greetings bot",
-      "intent": "Greeting",
-      "entities": []
-    },
-    {
-      "text": "hello",
-      "intent": "Greeting",
-      "entities": []
-    },
-    {
-      "text": "hi",
-      "intent": "Greeting",
-      "entities": []
-    },
-    {
-      "text": "hey there",
-      "intent": "Greeting",
-      "entities": []
-    }
-  ]
-}
