--- PictureBot/packages.config
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Autofac" version="3.5.2" targetFramework="net46" />
-  <package id="Chronic.Signed" version="0.3.2" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi" version="5.2.3" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net46" />
-  <package id="Microsoft.Azure.Search" version="3.0.3" targetFramework="net46" />
-  <package id="Microsoft.Bot.Builder" version="3.5.5" targetFramework="net46" />
-  <package id="Microsoft.IdentityModel.Protocol.Extensions" version="1.0.2.206221351" targetFramework="net46" />
-  <package id="Microsoft.Rest.ClientRuntime" version="2.3.4" targetFramework="net46" />
-  <package id="Microsoft.Rest.ClientRuntime.Azure" version="3.3.4" targetFramework="net46" />
-  <package id="Microsoft.Spatial" version="6.15.0" targetFramework="net46" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.1.0" targetFramework="net46" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net46" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="4.0.2.206221351" targetFramework="net46" />
-</packages>
